refactor : githubClient 인터페이스화를 위한 리팩터링
