+++ src/main/java/nextstep/member/application/GithubClient.java
@@ -1,55 +1,5 @@
 package nextstep.member.application;
 
-import nextstep.member.application.dto.GithubAccessTokenRequest;
-import nextstep.member.application.dto.GithubAccessTokenResponse;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.MediaType;
-import org.springframework.stereotype.Component;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.client.HttpClientErrorException;
-import org.springframework.web.client.RestTemplate;
-
-@Component
-public class GithubClient {
-
-    @Value("${github.client.id}")
-    private String clientId;
-    @Value("${github.client.secret}")
-    private String clientSecret;
-    @Value("${github.url.access-token}")
-    private String tokenUrl;
-
-    private RestTemplate restTemplate;
-
-    public GithubClient(RestTemplateBuilder restTemplateBuilder) {
-        this.restTemplate = restTemplateBuilder.build();
-    }
-
-    public String getAccessTokenFromGithub(String code) {
-        GithubAccessTokenRequest githubAccessTokenRequest = new GithubAccessTokenRequest(
-                code,
-                clientId,
-                clientSecret
-        );
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Accept", MediaType.APPLICATION_JSON_VALUE);
-
-        HttpEntity<MultiValueMap<String, String>> httpEntity = new HttpEntity(
-                githubAccessTokenRequest, headers);
-
-        String accessToken = restTemplate
-                .exchange(tokenUrl, HttpMethod.POST, httpEntity, GithubAccessTokenResponse.class)
-                .getBody()
-                .getAccessToken();
-        if (accessToken == null) {
-            throw new RuntimeException();
-        }
-        return accessToken;
-    }
+public interface GithubClient {
+    String getAccessTokenFromGithub(String code);
 }
-
